+++ src/impl_serde_db/de.rs
@@ -1,5 +1,5 @@
+use bigdecimal::ToPrimitive;
 use chrono::{NaiveDate, NaiveDateTime, NaiveTime};
-use num::cast::ToPrimitive;
 use serde_db::de::{
     ConversionError, DbValue, DbValueInto, DeserializableResultset, DeserializableRow,
     DeserializationError, DeserializationResult,
@@ -197,10 +197,9 @@ impl DbValueInto<u8> for TypedValue {
                     Err(number_range(i, "u8"))
                 }
             }
-            TypedValue::DECIMAL(dec) | TypedValue::N_DECIMAL(Some(dec)) => dec.into_decimal()
-                .to_u8()
-                .ok_or_else(|| decimal_range("u8")),
-
+            TypedValue::DECIMAL(bigdec) | TypedValue::N_DECIMAL(Some(bigdec)) => {
+                bigdec.to_u8().ok_or_else(|| decimal_range("u8"))
+            }
             value => Err(wrong_type(&value, "u8")),
         }
     }
@@ -233,10 +232,9 @@ impl DbValueInto<u16> for TypedValue {
                     Err(number_range(i, "u16"))
                 }
             }
-            TypedValue::DECIMAL(dec) | TypedValue::N_DECIMAL(Some(dec)) => dec.into_decimal()
-                .to_u16()
-                .ok_or_else(|| decimal_range("u16")),
-
+            TypedValue::DECIMAL(bigdec) | TypedValue::N_DECIMAL(Some(bigdec)) => {
+                bigdec.to_u16().ok_or_else(|| decimal_range("u16"))
+            }
             value => Err(wrong_type(&value, "u16")),
         }
     }
@@ -267,10 +265,9 @@ impl DbValueInto<u32> for TypedValue {
                     Err(number_range(i, "u32"))
                 }
             }
-            TypedValue::DECIMAL(dec) | TypedValue::N_DECIMAL(Some(dec)) => dec.into_decimal()
-                .to_u32()
-                .ok_or_else(|| decimal_range("u32")),
-
+            TypedValue::DECIMAL(bigdec) | TypedValue::N_DECIMAL(Some(bigdec)) => {
+                bigdec.to_u32().ok_or_else(|| decimal_range("u32"))
+            }
             value => Err(wrong_type(&value, "u32")),
         }
     }
@@ -300,9 +297,9 @@ impl DbValueInto<u64> for TypedValue {
                 Err(number_range(i, "u64"))
             },
 
-            TypedValue::DECIMAL(dec) | TypedValue::N_DECIMAL(Some(dec)) => dec.into_decimal()
-                .to_u64()
-                .ok_or_else(|| decimal_range("u64")),
+            TypedValue::DECIMAL(bigdec) | TypedValue::N_DECIMAL(Some(bigdec)) => {
+                bigdec.to_u64().ok_or_else(|| decimal_range("u64"))
+            }
             value => Err(wrong_type(&value, "u64")),
         }
     }
@@ -341,10 +338,9 @@ impl DbValueInto<i8> for TypedValue {
                     Err(number_range(i, "i8"))
                 }
             }
-            TypedValue::DECIMAL(dec) | TypedValue::N_DECIMAL(Some(dec)) => dec.into_decimal()
-                .to_i8()
-                .ok_or_else(|| decimal_range("i8")),
-
+            TypedValue::DECIMAL(bigdec) | TypedValue::N_DECIMAL(Some(bigdec)) => {
+                bigdec.to_i8().ok_or_else(|| decimal_range("i8"))
+            }
             value => Err(wrong_type(&value, "i8")),
         }
     }
@@ -373,10 +369,9 @@ impl DbValueInto<i16> for TypedValue {
                     Err(number_range(i, "i16"))
                 }
             }
-            TypedValue::DECIMAL(dec) | TypedValue::N_DECIMAL(Some(dec)) => dec.into_decimal()
-                .to_i16()
-                .ok_or_else(|| decimal_range("i16")),
-
+            TypedValue::DECIMAL(dec) | TypedValue::N_DECIMAL(Some(dec)) => {
+                dec.to_i16().ok_or_else(|| decimal_range("i16"))
+            }
             value => Err(wrong_type(&value, "i16")),
         }
     }
@@ -399,9 +394,9 @@ impl DbValueInto<i32> for TypedValue {
                     Err(number_range(i, "i32"))
                 }
             }
-            TypedValue::DECIMAL(dec) | TypedValue::N_DECIMAL(Some(dec)) => dec.into_decimal()
-                .to_i32()
-                .ok_or_else(|| decimal_range("i32")),
+            TypedValue::DECIMAL(bigdec) | TypedValue::N_DECIMAL(Some(bigdec)) => {
+                bigdec.to_i32().ok_or_else(|| decimal_range("i32"))
+            }
             value => Err(wrong_type(&value, "i32")),
         }
     }
@@ -416,9 +411,9 @@ impl DbValueInto<i64> for TypedValue {
             TypedValue::INT(i) | TypedValue::N_INT(Some(i)) => Ok(i64::from(i)),
             TypedValue::BIGINT(i) | TypedValue::N_BIGINT(Some(i)) => Ok(i),
             TypedValue::LONGDATE(ld) | TypedValue::N_LONGDATE(Some(ld)) => Ok(*ld.ref_raw()),
-            TypedValue::DECIMAL(dec) | TypedValue::N_DECIMAL(Some(dec)) => dec.into_decimal()
-                .to_i64()
-                .ok_or_else(|| decimal_range("i64")),
+            TypedValue::DECIMAL(bigdec) | TypedValue::N_DECIMAL(Some(bigdec)) => {
+                bigdec.to_i64().ok_or_else(|| decimal_range("i64"))
+            }
             value => Err(wrong_type(&value, "i64")),
         }
     }
@@ -428,9 +423,9 @@ impl DbValueInto<i64> for TypedValue {
 impl DbValueInto<f32> for TypedValue {
     fn try_into(self) -> Result<f32, ConversionError> {
         match self {
-            TypedValue::DECIMAL(dec) | TypedValue::N_DECIMAL(Some(dec)) => dec.into_decimal()
-                .to_f32()
-                .ok_or_else(|| decimal_range("f32")),
+            TypedValue::DECIMAL(bigdec) | TypedValue::N_DECIMAL(Some(bigdec)) => {
+                bigdec.to_f32().ok_or_else(|| decimal_range("f32"))
+            }
             TypedValue::REAL(f) | TypedValue::N_REAL(Some(f)) => Ok(f),
             value => Err(wrong_type(&value, "f32")),
         }
@@ -441,9 +436,9 @@ impl DbValueInto<f32> for TypedValue {
 impl DbValueInto<f64> for TypedValue {
     fn try_into(self) -> Result<f64, ConversionError> {
         match self {
-            TypedValue::DECIMAL(dec) | TypedValue::N_DECIMAL(Some(dec)) => dec.into_decimal()
-                .to_f64()
-                .ok_or_else(|| decimal_range("f64")),
+            TypedValue::DECIMAL(bigdec) | TypedValue::N_DECIMAL(Some(bigdec)) => {
+                bigdec.to_f64().ok_or_else(|| decimal_range("f64"))
+            }
             TypedValue::DOUBLE(f) | TypedValue::N_DOUBLE(Some(f)) => Ok(f),
             value => Err(wrong_type(&value, "f64")),
         }
@@ -489,7 +484,8 @@ impl DbValueInto<String> for TypedValue {
             TypedValue::CLOB(clob)
             | TypedValue::NCLOB(clob)
             | TypedValue::N_CLOB(Some(clob))
-            | TypedValue::N_NCLOB(Some(clob)) => Ok(clob.into_string()
+            | TypedValue::N_NCLOB(Some(clob)) => Ok(clob
+                .into_string()
                 .map_err(|e| ConversionError::Incomplete(e.description().to_owned()))?),
 
             value => Err(wrong_type(&value, "String")),
@@ -520,7 +516,8 @@ impl DbValueInto<NaiveDateTime> for TypedValue {
 impl DbValueInto<Vec<u8>> for TypedValue {
     fn try_into(self) -> Result<Vec<u8>, ConversionError> {
         match self {
-            TypedValue::BLOB(blob) | TypedValue::N_BLOB(Some(blob)) => Ok(blob.into_bytes()
+            TypedValue::BLOB(blob) | TypedValue::N_BLOB(Some(blob)) => Ok(blob
+                .into_bytes()
                 .map_err(|e| ConversionError::Incomplete(e.description().to_owned()))?),
 
             TypedValue::BINARY(v)