+++ tests/test_025_decimals.rs
@@ -2,46 +2,34 @@ extern crate flexi_logger;
 extern crate hdbconnect;
 #[macro_use]
 extern crate log;
+extern crate bigdecimal;
 extern crate num;
-extern crate rust_decimal;
 #[macro_use]
 extern crate serde_derive;
 extern crate serde_json;
 
 mod test_utils;
 
+use bigdecimal::BigDecimal;
 #[allow(unused_imports)]
 use flexi_logger::{LogSpecification, ReconfigurationHandle};
-use hdbconnect::{Connection, HdbResult};
+use hdbconnect::{Connection, HdbResult, HdbValue};
 use num::FromPrimitive;
-use rust_decimal::Decimal;
 
-#[test] // cargo test --test test_025_decimals -- --nocapture
-pub fn test_025_decimals() {
-    let mut log_handle = test_utils::init_logger("info");
-
-    match impl_test_025_decimals(&mut log_handle) {
-        Err(e) => {
-            error!("impl_test_025_decimals() failed with {:?}", e);
-            assert!(false)
-        }
-        Ok(_) => debug!("impl_test_025_decimals() ended successful"),
-    }
-}
-
-// Test the various ways to evaluate a resultset
-fn impl_test_025_decimals(log_handle: &mut ReconfigurationHandle) -> HdbResult<()> {
+//cargo test --test test_025_decimals -- --nocapture
+#[test]
+fn test_025_decimals() -> HdbResult<()> {
+    let mut log_handle = test_utils::init_logger("info, test_025_decimals = info");
     let mut connection = test_utils::get_authenticated_connection()?;
-    evaluate_resultset(log_handle, &mut connection)?;
+    test_025_decimals_impl(&mut log_handle, &mut connection)?;
     info!("{} calls to DB were executed", connection.get_call_count()?);
     Ok(())
 }
 
-fn evaluate_resultset(
+fn test_025_decimals_impl(
     _log_handle: &mut ReconfigurationHandle,
     connection: &mut Connection,
 ) -> HdbResult<()> {
-    // prepare the db table
     connection.multiple_statements_ignore_err(vec!["drop table TEST_DECIMALS"]);
     let stmts = vec![
         "create table TEST_DECIMALS (f1 NVARCHAR(100) primary key, f2 DECIMAL(7,5), f3 integer)",
@@ -61,14 +49,14 @@ fn evaluate_resultset(
         #[serde(rename = "F1")]
         f1: String,
         #[serde(rename = "F2")]
-        f2: Decimal,
+        f2: BigDecimal,
     };
 
     let insert_stmt_str = "insert into TEST_DECIMALS (F1, F2) values(?, ?)";
 
     // prepare & execute
     let mut insert_stmt = connection.prepare(insert_stmt_str)?;
-    insert_stmt.add_batch(&("75.53500", Decimal::from_f32(75.53500).unwrap()))?;
+    insert_stmt.add_batch(&("75.53500", BigDecimal::from_f32(75.53500).unwrap()))?;
     insert_stmt.add_batch(&("87.65432", 87.654_32_f32))?;
     insert_stmt.add_batch(&("0.00500", 0.005_f32))?;
     insert_stmt.add_batch(&("-0.00600", -0.006_00_f64))?;
@@ -82,25 +70,43 @@ fn evaluate_resultset(
     insert_stmt.add_batch(&("-8.65432", "-8.65432"))?;
     insert_stmt.execute_batch()?;
 
-    info!("Read and verify decimals to struct");
-    let resultset = connection.query("select f1, f2 from TEST_DECIMALS")?;
-    let result: Vec<TestData> = resultset.try_into()?;
-    for td in result {
-        debug!("{}, {}", td.f1, td.f2);
-        assert_eq!(td.f1, format!("{}", td.f2));
+    info!("Read and verify decimals");
+    let resultset = connection.query("select f1, f2 from TEST_DECIMALS order by f2")?;
+    let precision = resultset.metadata().precision(1)?;
+    let scale = resultset.metadata().scale(1)? as usize;
+    for row in resultset {
+        let row = row?;
+        let f1: HdbValue = row.cloned_value(0)?;
+        let f2: HdbValue = row.cloned_value(1)?;
+        if let HdbValue::N_DECIMAL(Some(ref bd)) = f2 {
+            debug!("precision = {}, scale = {}", precision, scale);
+            assert_eq!(format!("{}", f1), format!("{0:.1$}", bd, scale));
+        } else {
+            assert!(false, "Unexpected value type");
+        }
     }
 
-    info!("Read and verify decimals to tuple");
-    let result: Vec<(String, String)> =
-        connection.query("select * from TEST_DECIMALS")?.try_into()?;
-    for row in result {
-        debug!("{}, {}", row.0, row.1);
-        assert_eq!(row.0, row.1);
+    info!("Read and verify decimals to struct");
+    let resultset = connection.query("select f1, f2 from TEST_DECIMALS order by f2")?;
+    let scale = resultset.metadata().scale(1)? as usize;
+    let result: Vec<TestData> = resultset.try_into()?;
+    for td in result {
+        debug!("{:?}, {:?}", td.f1, td.f2);
+        assert_eq!(td.f1, format!("{}", format!("{0:.1$}", td.f2, scale)));
     }
+
+    // Does not work because the scale information is not available
+    // info!("Read and verify decimals to tuple");
+    // let result: Vec<(String, String)> =
+    //     connection.query("select * from TEST_DECIMALS")?.try_into()?;
+    // for row in result {
+    //     debug!("{}, {}", row.0, row.1);
+    //     assert_eq!(row.0, row.1);
+    // }
 
     info!("Read and verify decimal to single value");
     let resultset = connection.query("select AVG(F3) from TEST_DECIMALS")?;
-    let mydata: Option<Decimal> = resultset.try_into()?;
+    let mydata: Option<BigDecimal> = resultset.try_into()?;
     assert_eq!(mydata, None);
 
     let mydata: Option<i64> = connection